+++ src/app/layout.tsx
@@ -9,8 +9,8 @@ import Navbar from "@/components/Common/Navbar/Navbar";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Agrify Carbon",
-  description: "Generated by create next app",
+  title: "Agrify Marketplace",
+  description: "Agrify Marketplace focuses on regenerative farming and carbon offsetting.",
 };
 
 export default function RootLayout({
