+++ src/app/page.tsx
@@ -1,6 +1,7 @@
 "use client";
 
-import ErrorPage from "@/components/ErrorPage/ErrorPage";
+import ErrorPage from "@/components/ErrorPageComponents/ErrorPage";
+
 
 
 export default function Login() {
