+++ src/components/Common/FourColumnTableRow/types.ts
