+++ src/components/Common/ReceiptModal/ReceiptModal.tsx
