+++ src/components/Common/ReceiptModal/types.ts
